Remove hive shim code
